# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: Multi Tenant SaaS API
  version: 1.3.0
  description: REST API for a multi-tenant SaaS platform.
servers:
  - url: /api
    description: API base path
...
paths:
...
  /admin/settings:
    get:
      tags: [Admin]
      summary: Get platform settings (super-admin)
      security:
        - sanctum: []
      responses:
        '200':
          description: Platform settings
        '403':
          description: Requires super-admin role
    put:
      tags: [Admin]
      summary: Update platform settings (super-admin)
      security:
        - sanctum: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings:
                  type: object
                  properties:
                    maintenance_mode: { type: boolean }
                    default_tenant_status: { type: string, enum: [active, inactive] }
                    max_users_per_tenant: { type: integer }
                    support_email: { type: string, format: email }
      responses:
        '200':
          description: Settings updated
        '403':
          description: Requires super-admin role
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: Obtain via POST /auth/login or POST /auth/register.
...